Fix warnings

+++ Pod/Server/NSData+SHA1.m
@@ -25,7 +25,7 @@
 {
     uint8_t digest[CC_SHA1_DIGEST_LENGTH];
     
-    CC_SHA1(self.bytes, self.length, digest);
+    CC_SHA1(self.bytes, (unsigned int)self.length, digest);
     
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_SHA1_DIGEST_LENGTH * 2];
